+++ CHANGES.md
@@ -4,6 +4,7 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
+- 279: when colleague has multiple roles on 1 assignment, show as 1 line, multiple chips
 - 281: fix label delete bug introduced in 266
 - 266: (migration) enforce uniqueness on Colleague.email + Colleague.source
 - 266: always create Colleague for Users